Share structs between orbuculum.c and serialFeeder_win32.c
